requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.25.0
contourpy==1.1.1
croniter==1.4.1
cycler==0.11.0
dateutils==0.6.12
deepdiff==6.5.0
docker-pycreds==0.4.0
exceptiongroup==1.1.3
fast-tmfg==0.0.8
fastapi==0.103.1
filelock==3.9.0
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
idna==3.4
importlib-resources==6.1.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lightning==2.0.9
lightning-cloud==0.5.38
lightning-utilities==0.9.0
lit==15.0.7
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numpy==1.26.0
ordered-set==4.1.0
packaging==23.1
pandas==2.1.1
pathtools==0.1.2
Pillow==10.0.1
polars==0.19.5
protobuf==4.24.3
psutil==5.9.5
pyarrow==13.0.0
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.9
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
requests==2.31.0
rich==13.5.3
scikit-learn==1.3.1
scipy==1.11.3
sentry-sdk==1.31.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
threadpoolctl==3.2.0
torch==2.0.0+cu118
torchinfo==1.8.0
torchmetrics==1.2.0
tqdm==4.66.1
traitlets==5.10.1
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
wandb==0.15.11
wcwidth==0.2.6
websocket-client==1.6.3
websockets==11.0.3
yarl==1.9.2
zipp==3.17.0
